+++ core/src/org/labkey/core/admin/rConfiguration.jsp
@@ -19,7 +19,7 @@
 <%@ page import="org.labkey.api.data.Container" %>
 <%@ page import="org.labkey.api.reports.ExternalScriptEngineDefinition" %>
 <%@ page import="org.labkey.api.reports.LabKeyScriptEngineManager" %>
-<%@ page import="org.labkey.api.reports.RemoteRNotEnabledException" %>
+<%@ page import="org.labkey.api.reports.report.r.RemoteRNotEnabledException" %>
 <%@ page import="org.labkey.api.services.ServiceRegistry" %>
 <%@ page import="org.labkey.api.util.element.Option.OptionBuilder" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
