+++ query/src/org/labkey/query/reports/ReportsController.java
@@ -71,8 +71,8 @@ import org.labkey.api.premium.PremiumService;
 import org.labkey.api.query.SimpleValidationError;
 import org.labkey.api.query.ValidationError;
 import org.labkey.api.query.ValidationException;
-import org.labkey.api.reports.RConnectionHolder;
-import org.labkey.api.reports.RemoteRNotEnabledException;
+import org.labkey.api.reports.report.r.RConnectionHolder;
+import org.labkey.api.reports.report.r.RemoteRNotEnabledException;
 import org.labkey.api.reports.Report;
 import org.labkey.api.reports.ReportContentEmailManager;
 import org.labkey.api.reports.ReportService;
